pg-az-backup/requirements.txt
requests
azure-storage-blob